5643 Move IteratorFactory and IteratorConfig to iterator package
